Restaurant configs implementation

+++ gastromanager-backend/src/main/java/com/kaiho/gastromanager/domain/ingredient/api/IngredientServicePort.java
@@ -21,7 +21,7 @@ public interface IngredientServicePort {
 
     List<Ingredient> getIngredientsByUuid(Set<UUID> uuids);
 
-    void batchAdjustStock(Map<UUID, Integer> stockAdjustments, String orderPlacement);
+    void batchAdjustStock(Map<UUID, Double> stockAdjustments, String orderPlacement);
 
     Restaurant getRestaurantByIngredientUuid(UUID ingredientUuid);
 }
